bottom sheet, exercise mistake view, theme change, design bug fixes and others

+++ app/src/main/java/com/example/studita/utils/NumberExtensions.kt
@@ -3,10 +3,10 @@ package com.example.studita.utils
 import android.content.res.Resources
 import java.math.BigDecimal
 
-fun Int.pxToDp(): Int = (this / Resources.getSystem().displayMetrics.density).toInt()
-fun Int.dpToPx(): Int = (this * Resources.getSystem().displayMetrics.density).toInt()
-fun Int.pxToSp(): Int = (this / Resources.getSystem().displayMetrics.scaledDensity).toInt()
-fun Int.spToPx(): Int = (this * Resources.getSystem().displayMetrics.scaledDensity).toInt()
+fun Int.pxToDp(): Float = (this / Resources.getSystem().displayMetrics.density)
+fun Float.dpToPx(): Int = (this * Resources.getSystem().displayMetrics.density).toInt()
+fun Int.pxToSp(): Float = (this / Resources.getSystem().displayMetrics.scaledDensity)
+fun Float.spToPx(): Int = (this * Resources.getSystem().displayMetrics.scaledDensity).toInt()
 
 fun Float.round(decimalPlace: Int): Float {
     var bd = BigDecimal(this.toString())
